requirements.txt
supabase>=2.0.0
composio-openai
composio-openai-agents>=0.8.0
openai>=1.50.0
openai-agents
python-dotenv
flask
flask-cors>=4.0.0
pydantic>=2.0.0
requests


fastapi>=0.104.0
uvicorn>=0.24.0

pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
botocore>=1.40.0
boto3>=1.40.0
